perf: Split out the back button into it's own file and memoize it
